Migrate all mission_control tests
