+++ mission_control/test/test_files/test_missions/mission.xml
@@ -3,7 +3,7 @@
 <root main_tree_to_execute = "MainTree" >
   <BehaviorTree ID="MainTree">
     <Sequence name="mission test">
-      <FixRudder depth="1.0" rudder="2.0" speed_knots="3.0" rudder_tol="5.0" depth_tol="6.0" altitude_tol="7.0"/>
+      <FixRudder depth="1.0" rudder="2.0" speed_knots="3.0"/>
     </Sequence>
   </BehaviorTree>
 </root>
