+++ mission_control/test/test_mission_control_behaviors.test
@@ -2,16 +2,18 @@
   <include file="$(find mission_control)/launch/mission_control.launch" />
   <!--
       These test verifyes the behaviors/actions:
-        -   Fixed Rudder
-        -   Depth Heading
-        -   Attitude Servo
-        -   Altitude Heading
+        -   FixRudder
+        -   SetDepthHeading
+        -   AttitudeServo
+        -   SetAltitudeHeading
+        -   GoToWaypoint
   -->
 
-  <test test-name="test_mission_control_fixed_rudder_action" pkg="mission_control" type="test_fixed_rudder_behavior.py" time-limit="10.0" />
-  <test test-name="test_mission_control_attitude_servo_action" pkg="mission_control" type="test_attitude_servo_behavior.py" time-limit="10.0" />
-  <test test-name="test_mission_control_depth_heading_action" pkg="mission_control" type="test_depth_heading_behavior.py" time-limit="10.0" />
-  <test test-name="test_mission_control_altitude_heading_action" pkg="mission_control" type="test_altitude_heading_behavior.py" time-limit="10.0" />
+  <test test-name="test_fix_rudder_action" pkg="mission_control" type="test_fix_rudder_action.py" time-limit="10.0" />
+  <test test-name="test_attitude_servo_action" pkg="mission_control" type="test_attitude_servo_action.py" time-limit="10.0" />
+  <test test-name="test_set_depth_heading_action" pkg="mission_control" type="test_set_depth_heading_action.py" time-limit="10.0" />
+  <test test-name="test_set_altitude_heading_action" pkg="mission_control" type="test_set_altitude_heading_action.py" time-limit="10.0" />
+  <test test-name="test_go_to_waypoint_action" pkg="mission_control" type="test_go_to_waypoint_action.py" time-limit="10.0" />
 
 </launch>
