+++ mission_control/test/test_set_depth_heading_action.py
@@ -40,69 +40,68 @@ import rosnode
 import rospy
 import rostest
 from mission_control.msg import ReportExecuteMissionState
-from mission_control.msg import AltitudeHeading
+from mission_control.msg import DepthHeading
 from auv_interfaces.msg import StateStamped
 from mission_interface import MissionInterface
 from mission_interface import wait_for
 
 
-class TestAltitudeHeadingBehavior(unittest.TestCase):
-    """
-        The test loads and execute an altitude heading behavior.
-        -   Load the mission altitude_heading_mission_test.xml
-        -   Execute the mission
-        -   Check if the behavior publishes the goal
-        -   Simulate auv_inteface data to finish the behavior
-        -   Test if the mission is COMPLETE
-    """
+class TestSetDepthHeadingAction(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        rospy.init_node('test_altitude_heading_behavior')
+        rospy.init_node('test_set_depth_heading')
 
     def setUp(self):
-        self.altitude_heading_goal = AltitudeHeading()
+        self.depth_heading_goal = DepthHeading()
         self.mission = MissionInterface()
 
-        self.altitude_heading_msg = rospy.Subscriber(
-            '/mngr/altitude_heading',
-            AltitudeHeading,
-            self.altitude_heading_goal_callback)
+        self.depth_heading_msg = rospy.Subscriber(
+            '/mngr/depth_heading',
+            DepthHeading,
+            self.depth_heading_goal_callback)
 
         self.simulated_auv_interface_data_pub = rospy.Publisher(
             '/state',
             StateStamped,
             queue_size=1)
 
-    def altitude_heading_goal_callback(self, msg):
-        self.altitude_heading_goal = msg
+    def depth_heading_goal_callback(self, msg):
+        self.depth_heading_goal = msg
 
-    def test_mission_with_altitude_heading_behavior(self):
-        self.mission.load_mission('altitude_heading_mission_test.xml')
+    def test_mission_with_set_depth_heading_action(self):
+        """
+        The test:
+        -  Loads a mission with a single SetDepthHeading action.
+        -  Executes the mission.
+        -  Waits until the action publishes a DepthHeading setpoint.
+        -  Simulates state updates for the action to complete.
+        -  waits until the mission is COMPLETE.
+        """
+        self.mission.load_mission('depth_heading_mission_test.xml')
         self.mission.execute_mission()
-        self.mission.read_behavior_parameters('AltitudeHeadingBehavior')
-        altitude = self.mission.get_behavior_parameter('altitude')
+
+        self.mission.read_behavior_parameters('SetDepthHeading')
+        depth = self.mission.get_behavior_parameter('depth')
         heading = self.mission.get_behavior_parameter('heading')
         speed_knots = self.mission.get_behavior_parameter('speed_knots')
 
         # Calculate the mask
         enable_mask = 0
-        if altitude is not None:
-            enable_mask |= AltitudeHeading.ALTITUDE_ENA
-
+        if depth is not None:
+            enable_mask |= DepthHeading.DEPTH_ENA
         if heading is not None:
-            enable_mask |= AltitudeHeading.HEADING_ENA
-
+            enable_mask |= DepthHeading.HEADING_ENA
         if speed_knots is not None:
-            enable_mask |= AltitudeHeading.SPEED_KNOTS_ENA
+            enable_mask |= DepthHeading.SPEED_KNOTS_ENA
 
-        def altitude_heading_goals_are_set():
-            return ((altitude is None or self.altitude_heading_goal.altitude == float(altitude)) and
-                    (heading is None or self.altitude_heading_goal.heading == float(heading)) and
-                    (speed_knots is None or self.altitude_heading_goal.speed_knots == float(speed_knots)) and
-                    self.altitude_heading_goal.ena_mask == enable_mask)
+        def depth_heading_goals_are_set():
+            return ((depth is None or self.depth_heading_goal.depth == float(depth)) and
+                    (heading is None or self.depth_heading_goal.heading == float(heading)) and
+                    (speed_knots is None or self.depth_heading_goal.speed_knots == float(speed_knots)) and
+                    self.depth_heading_goal.ena_mask == enable_mask)
 
-        self.assertTrue(wait_for(altitude_heading_goals_are_set),
+        self.assertTrue(wait_for(depth_heading_goals_are_set),
                         msg='Mission control must publish goals')
 
         # send data to finish the mission
@@ -119,6 +118,5 @@ class TestAltitudeHeadingBehavior(unittest.TestCase):
                         msg='Mission control must report only COMPLETE')
 
 
-if __name__ == "__main__":
-    rostest.rosrun('mission_control', 'mission_control_test_altitude_heading_behavior',
-                   TestAltitudeHeadingBehavior)
+if __name__ == '__main__':
+    rostest.rosrun('mission_control', 'test_set_depth_heading_action', TestSetDepthHeadingAction)
